en principio todo funciona

+++ src/main/java/com/example/DB1Crud/Servicios/StudentServicio.java
@@ -3,6 +3,8 @@ package com.example.DB1Crud.Servicios;
 import com.example.DB1Crud.DTOs.input.StudentInputDTO;
 import com.example.DB1Crud.Excepciones.NotFoundException;
 import com.example.DB1Crud.POJOs.Student;
+import com.example.DB1Crud.Repositorios.PersonaRepositorio;
+import com.example.DB1Crud.Repositorios.ProfesorRepositorio;
 import com.example.DB1Crud.Repositorios.StudentRepositorio;
 import com.example.DB1Crud.Excepciones.UnprocessableEntityException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +18,12 @@ import java.util.Optional;
 public class StudentServicio {
     @Autowired
     StudentRepositorio studentRepositorio;
+
+    @Autowired
+    ProfesorRepositorio profesorRepositorio;
+
+    @Autowired
+    PersonaRepositorio personaRepositorio;
 
     public Optional<Student> buscarId(int id_student){
         Optional<Student> estudiante = studentRepositorio.findById(id_student);
@@ -34,6 +42,8 @@ public class StudentServicio {
         if(student==null){
             throw new UnprocessableEntityException("El estudiante es nulo");
         }else{
+            student.setPersonaRepositorio(this.personaRepositorio);
+            student.setProfesorRepositorio(this.profesorRepositorio); //de lokos si funciona
             Student st = new Student(student);
             studentRepositorio.save(st);
             System.out.println("Estudiante guardado");
@@ -41,11 +51,27 @@ public class StudentServicio {
     }
 
     public void delete(int id_student) {
-        studentRepositorio.deleteById(id_student);
+        Optional<Student> s;
+        s = buscarId(id_student);
+        if(s==null){
+            throw new NotFoundException("No hay usuarios");
+        }else{
+            studentRepositorio.deleteById(id_student);
+        }
     }
 
     public void updateStudent(int id_student, StudentInputDTO s) throws Exception{
-        addStudent(s);
-        delete(id_student);
+
+        Optional<Student> aux;
+        aux = studentRepositorio.findById(id_student);
+        aux.get().setProfesor(profesorRepositorio.findById(s.getId_profesor()).get());
+        aux.get().setPersona(personaRepositorio.findById(s.getId_persona()).get());
+        aux.get().setId_student(id_student);
+        aux.get().setBranch(s.getBranch());
+        aux.get().setComents(s.getComents());
+        aux.get().setEstudios(s.getEstudios());
+        aux.get().setNum_hours_week(s.getNum_hours_week());
+
+        studentRepositorio.save(aux.get());
     }
 }
